% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chisq_pearson}
\alias{chisq_pearson}
\title{Compute the Pearson Chi-Squared statistic.}
\arguments{
\item{\code{d}}{The observed vector}

\item{\code{e}}{The expected vector}
}
\value{
The computed value of the statistic
}
\description{
Compute the Pearson Chi-Squared statistic.
}
