% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{copy_sql_to_sql}
\alias{copy_sql_to_sql}
\title{Copy SQL table}
\usage{
copy_sql_to_sql(
  in_conn,
  in_table,
  out_conn,
  out_table,
  initial_rows = 10000L,
  batch_bytes = getOption("ffbatchbytes")
)
}
\arguments{
\item{\code{in_conn}}{database to copy from (DBI connection)}

\item{\code{in_table}}{name of table to copy from}

\item{\code{out_conn}}{database to copy to (DBI connection)}

\item{\code{out_table}}{name of table to copy to}

\item{\code{initial_rows}}{number of rows to read in first batch}

\item{\code{batch_bytes}}{size of subsequent batches in bytes}
}
\description{
Copy a table from one SQL database to another.
}
\details{
The copy is performed by fetching the data from the input DB to
memory, then pushing it from memory to the output DB. To avoid excessive
memory usage, the table is read in batches.

Warning: if the output table already exists, it will be dropped!
}
