% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{_graph_density}
\alias{_graph_density}
\alias{graph_density}
\title{Compute a value of the density of the links between the values in each
group.}
\usage{
graph_density(values, groups)
}
\arguments{
\item{\code{values}}{The vector of values}

\item{\code{groups}}{The vector of groups, corresponding to the values}
}
\value{
A numeric vector of the values for each group, named by group.
}
\description{
Compute a value of the density of the links between the values in each
group.
}
