% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{group}
\alias{group}
\title{Group a vector of values, using a corresponding group vector.}
\arguments{
\item{\code{value_col}}{The vector of values}

\item{\code{group_col}}{The vector of groups, corresponding to the values}
}
\value{
A map from group -> vector of values
}
\description{
Group a vector of values, using a corresponding group vector.
}
