% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_env_backends.R
\name{lazy_fread}
\alias{lazy_fread}
\title{Lazy environment from flat file}
\usage{
lazy_fread(input, ...)
}
\description{
The objects in the environment are the columns of the file. Supports all
options of \code{fread} package in \code{data.table} package.
}
