% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_group_count}
\alias{max_group_count}
\title{Find the largest count in a group.}
\arguments{
\item{\code{group}}{A map from the counts of the group}
}
\value{
The size of the group
}
\description{
Find the largest count in a group.
}
