% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{run_model}
\alias{run_model}
\title{Run the model}
\usage{
run_model(
  input,
  input_table = NULL,
  input_stats = TRUE,
  output = NULL,
  create_indices = TRUE,
  store_input = FALSE,
  env = NULL
)
}
\arguments{
\item{\code{input}}{input data source, of one of the following types:
\itemize{
 \item data frame
 \item environment
 \item filename (supported file types: CSV, TSV)
 \item SQL database (DBI connection)
}}

\item{\code{input_table}}{name of table containing input data.
Only applicable when input source is a SQL database.}

\item{\code{input_stats}}{whether to compute summary statistics for input
data columns (default: yes)}

\item{\code{output}}{output database (DBI connection, optional)}

\item{\code{create_indices}}{whether to create group indices in the 
entity-level tables of the output DB (default: yes)}

\item{\code{store_input}}{whether to save the input data as an additional
table in the output DB (default: no)}

\item{\code{env}}{environment containing the model configuration
(default: the global model environment)}
}
\value{
A named list with the following data frames:
\enumerate{
 \item \code{run_summary}:
   A table of key-value pairs summarizing the run.

 \item \code{entity_metric_values}:
   Entity-level metric values. Group metrics are not included.

 \item \code{entity_metric_stats}:
   Population statistics of entity-level metric values. The columns are
   metrics and the rows are summary statistics.
   
   No statistics are computed for group metrics.
 
 \item \code{group_attributes}:
   General attributes of the group. At present, this only includes the size
   of the group.
 
 \item \code{group_metric_values}:
   Group-level metric values. For non-group metrics, these are simply
   the means of the entity-level metric values.
 
 \item \code{group_metric_scores}:
   Group-level metric scores
 
 \item \code{group_composite_scores}:
   Group-level composite scores
 
 \item \code{input_stats}:
   Population statistics for input data. The columns are inputs and the rows
   are summary statistics.
}
The tables \code{run_summary}, \code{input_stats}, and
\code{entity_metric_stats} have type \code{data.table}.
All other tables have type \code{ffdf}.
}
\description{
Runs the model on the specified input data.
}
