% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{_uniq_disc}
\alias{_uniq_disc}
\alias{uniq_disc}
\title{Compute a ratio of the discrete unique values in each group.}
\usage{
uniq_disc(values, groups, type)
}
\arguments{
\item{\code{values}}{The vector of values}

\item{\code{groups}}{The vector of groups, corresponding to the values}

\item{\code{type}}{The type of ratio to compute: "distinct" or "frequent"}
}
\value{
A numeric vector of the values for each group, named by group.
}
\description{
Compute a ratio of the discrete unique values in each group.
}
